+++ src/__mocks__/vscode.ts
@@ -1,3 +1,7 @@
+export const commands = {
+  executeCommand: jest.fn(),
+}
+
 export const workspace = {
   getConfiguration: jest.fn(() => new Map()),
   onDidChangeConfiguration: jest.fn(),